job-search-agent/pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
]


